refactor: split generative models
